permisions for console

+++ server/src/server/background/ChatServer.ts
@@ -25,44 +25,49 @@ export class ChatServer {
         this.time = new Date();
         // TODO: check for privileges here!!!
 
-        const message = consoleMessage.message.split(" ");
+        for(let i = 0; i < gameServer.admins.length; i++) {
+            if(consoleMessage.username == gameServer.admins[i]) {
 
-        switch(message[0]) {
-            case "/i":
-                switch(message[1]) {
-                    case "e":
-                        const entity = await gameServer.getEntityByUUID(message[2]);
-                        this._sendConsoleMessageToAll(JSON.stringify(entity));
+                const message = consoleMessage.message.split(" ");
+
+                switch(message[0]) {
+                    case "/i":
+                        switch(message[1]) {
+                            case "e":
+                                const entity = await gameServer.getEntityByUUID(message[2]);
+                                this._sendConsoleMessageToAll(JSON.stringify(entity));
+                                break;
+                            case "p":
+                                const player = await gameServer.getPlayerByUsername(message[2]);
+                                this._sendConsoleMessageToAll(JSON.stringify(player));
+                                break;
+                        }
                         break;
-                    case "p":
-                        const player = await gameServer.getPlayerByUsername(message[2]);
-                        this._sendConsoleMessageToAll(JSON.stringify(player));
+                    case "/c":
+                        switch(message[1]) {
+                            case "e": 
+                                const player = await gameServer.getPlayerByUsername(consoleMessage.username);
+                                if(player?.currentMap) {
+                                    gameServer.spacemaps[player.currentMap].spawnAlien(message[2], {
+                                        x: (0.5 - Math.random()) * 10,
+                                        y: (0.5 - Math.random()) * 10,
+                                    });
+                                }
+                                break;
+                        }
                         break;
-                }
-                break;
-            case "/c":
-                switch(message[1]) {
-                    case "e": 
-                        const player = await gameServer.getPlayerByUsername(consoleMessage.username);
-                        if(player?.currentMap) {
-                            gameServer.spacemaps[player.currentMap].spawnAlien(message[2], {
-                                x: (0.5 - Math.random()) * 10,
-                                y: (0.5 - Math.random()) * 10,
-                            });
+                    case "/d":
+                        switch(message[1]) {
+                            case "e":
+                                const player = await gameServer.getPlayerByUsername(consoleMessage.username);
+                                if(player?.currentMap) {
+                                    gameServer.spacemaps[player.currentMap].deleteAlienByuuid(message[2]);
+                                }
+                                break;
                         }
                         break;
                 }
-                break;
-            case "/d":
-                switch(message[1]) {
-                    case "e":
-                        const player = await gameServer.getPlayerByUsername(consoleMessage.username);
-                        if(player?.currentMap) {
-                            gameServer.spacemaps[player.currentMap].deleteAlienByuuid(message[2]);
-                        }
-                        break;
-                }
-                break;
+            }
         }
 
         // TODO: main logic
